Enhances skills section and adds contact form
Improves the skills section by adding more skills and categories.
The title now has a more appealing style and a subtitle to provide
more context.

Adds a contact page with a functional form. The form integrates
with an external service to send messages and provides user
feedback upon submission. Also, the browser language is detected
to set the default language.